fixed demo tick timing, team numbers, nav aabb index

+++ analytics/include/queries/train_dataset.h
@@ -21,7 +21,7 @@ public:
     };
 
     struct TimeStepState {
-        uint32_t curAABB;
+        size_t curAABB;
         Vec3 pos;
         vector<NavmeshState> navStates;
         // these aren't printed, just used for bookkeeping during query
